Add SecurePreferences interface
