+++ app/src/community/java/com/twilio/video/app/auth/CommunityAuthModule.kt
@@ -19,8 +19,8 @@ import android.content.SharedPreferences
 import com.twilio.video.app.ApplicationScope
 import com.twilio.video.app.data.AuthServiceModule
 import com.twilio.video.app.data.api.TokenService
+import com.twilio.video.app.security.SecurePreferences
 import com.twilio.video.app.security.SecurityModule
-import com.twilio.video.app.util.SecurePreferences
 import dagger.Module
 import dagger.Provides
