+++ app/src/community/java/com/twilio/video/app/auth/CommunityAuthenticator.kt
@@ -22,7 +22,7 @@ import com.twilio.video.app.auth.LoginResult.CommunityLoginSuccessResult
 import com.twilio.video.app.data.PASSCODE
 import com.twilio.video.app.data.Preferences.DISPLAY_NAME
 import com.twilio.video.app.data.api.TokenService
-import com.twilio.video.app.util.SecurePreferences
+import com.twilio.video.app.security.SecurePreferences
 import com.twilio.video.app.util.putString
 import com.twilio.video.app.util.remove
 import io.reactivex.Observable
@@ -49,7 +49,7 @@ class CommunityAuthenticator constructor(
                     tokenService.getToken(identity = loginEvent.identity, passcode = loginEvent.passcode)
 
                     sharedPreferences.putString(DISPLAY_NAME, loginEvent.identity)
-                    securePreferences.putString(PASSCODE, loginEvent.passcode)
+                    securePreferences.putSecureString(PASSCODE, loginEvent.passcode)
 
                     CommunityLoginSuccessResult
                 } catch (e: Exception) {
@@ -63,7 +63,7 @@ class CommunityAuthenticator constructor(
     }
 
     override fun loggedIn(): Boolean {
-        return !securePreferences.getString(PASSCODE).isNullOrEmpty()
+        return !securePreferences.getSecureString(PASSCODE).isNullOrEmpty()
     }
 
     override fun logout() {
